move config.ini to config.example to make upgrades easier. (#163)
* move config.ini to config.example to make upgrades easier.

Update .gitignore and README.md files

* Add phase information for OpenDTU and ahoy

* Add check for config.ini file existence

* Add AcPosition information for OpenDTU in README.md

* Change from `mv` to `cp`

* add newline to README.md

+++ install.sh
@@ -2,10 +2,18 @@
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 SERVICE_NAME=$(basename $SCRIPT_DIR)
 
+
+# check if config.ini file exists
+if [ ! -f $SCRIPT_DIR/config.ini ]; then
+    echo "config.ini file not found. Please make sure it exists. If not created yet, please copy it from config.example."
+    exit 1
+fi
+
 # delete old logs if they exist  
 if [ -f /data/dbus-opendtu/current.log ]; then  
     rm /data/dbus-opendtu/current.log*  
 fi 
+
 
 # set permissions for script files
 chmod a+x $SCRIPT_DIR/restart.sh
